requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==4.2.22
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0
